Separted Navbar render logic.
